fixed everything. it all works. making a pull request
