--- experiments/notebooks/utils/policy_selector_widget/__init__.py
@@ -1,3 +1,0 @@
-from .policy_selector_widget import PolicySelectorWidget, create_policy_selector_widget
-
-__all__ = ["PolicySelectorWidget", "create_policy_selector_widget"]
