+++ experiments/notebooks/utils/scorecard_widget/package.json
@@ -12,9 +12,13 @@
     "vite-plugin-node-polyfills": "^0.24.0"
   },
   "dependencies": {
+    "object-assign": "^4.1.1",
     "plotly.js": "^3.0.3",
+    "prop-types": "^15.8.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-plotly.js": "^2.6.0"
+    "react-is": "^19.1.1",
+    "react-plotly.js": "^2.6.0",
+    "scheduler": "^0.26.0"
   }
 }
